refactor: extract COmmandItem to own file
